requirements.txt
amqp==5.0.5
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.44.0
django-filter==2.4.0
django-modeltranslation==0.16.2
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
easy-thumbnails==2.7.1
gunicorn==20.0.4
idna==2.10
jsmin==2.2.2
kombu==5.0.2
oauthlib==3.1.0
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
Wand==0.6.5
wcwidth==0.2.5
